api_code/requirements.txt
amqp==2.6.1
asgiref==3.4.1
billiard==3.6.4.0
CacheControl==0.12.11
celery==4.4.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.7
django-audiofield @ git+https://github.com/romonzaman/django-audiofield.git@9108921d8816ec19a826799dca72a5af09da4c51
django-celery==3.1.17
django-extensions==3.1.5
django-filter==2.4.0
djangorestframework==3.12.4
flowroutenumbersandmessaging @ git+https://github.com/flowroute/flowroute-sdk-v3-python.git
FreeSWITCH-ESL-Python==1.2
idna==3.3
importlib-metadata==4.8.1
jsonpickle==0.9.6
kombu==4.6.11
Markdown==3.3.4
msgpack==1.0.4
prompt-toolkit==3.0.29
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2021.1
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.9
vine==1.3.0
wcwidth==0.2.5
zipp==3.5.0
plivo==4.24.1
django-admin-tools==0.9.2
